% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-util.R
\name{getFXdata}
\alias{getFXdata}
\title{Get financial data from the Alpha Vantage API}
\usage{
getFXdata(FUN, apiKey, from_symbol = NULL, to_symbol = NULL,
  interval = NULL, outputsize = "full", datatype = "csv", ...)
}
\arguments{
\item{FUN}{function list that matches with Aplphavantag function list}

\item{from_symbol}{base currency}

\item{to_symbol}{quote currency}

\item{interval}{frequency of the data}

\item{outputsize}{compact or full}

\item{datatype}{csv or json}
}
\value{
returns the time series OHLC
}
\description{
Get financial data from the Alpha Vantage API
}
