% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-util.R
\name{multipleFX_pair}
\alias{multipleFX_pair}
\alias{multipleFX_Pair}
\title{Get financial data from the Alpha Vantage API using multiple currency pairs}
\usage{
multipleFX_Pair(currencyPair, apiKey = NULL, ...)
}
\arguments{
\item{currencyPair}{currency pairs of our interest}
}
\value{
returns the time series OHLC in a list
}
\description{
Get financial data from the Alpha Vantage API using multiple currency pairs
}
